:sparkles: Feat: subscription confirm email
Subscription confirm email added

+++ app/Repositories/Frontend/SubscriptionRepository.php
@@ -2,8 +2,9 @@
 
 namespace App\Repositories\Frontend;
 
-use Illuminate\Http\Request;
 use App\Models\Subscription;
+use App\Exceptions\AlreadyUnsubscribedException;
+use App\Exceptions\MissingSubscriptionException;
 
 /**
  * Subscription entity database query class.
@@ -13,15 +14,37 @@ use App\Models\Subscription;
 class SubscriptionRepository
 {
     /**
-     * Delete subscription instance from the database.
+     * Update subscription instance in the database.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  string  $email
+     * @param  string  $remember_token
      * @return void
      */
-    public static function unsubscribe(Request $request)
+    public static function confirm($email, $remember_token)
     {
-        $subscription = Subscription::where('remember_token', $request->remember_token)
-                ->firstOrFail();
+        $subscription = Subscription::where('email', $email)
+                ->where('remember_token', $remember_token)
+                ->first();
+        if (!$subscription) {
+            throw new MissingSubscriptionException('Missing email');
+        }
+        $subscription->subscribed = 1;
+        $subscription->save();
+    }
+
+    /**
+     * Delete subscription instance from the database.
+     *
+     * @param  string  $remember_token
+     * @return void
+     */
+    public static function unsubscribe($remember_token)
+    {
+        $subscription = Subscription::where('remember_token', $remember_token)
+                ->first();
+        if (!$subscription) {
+            throw new AlreadyUnsubscribedException('Email has been already deleted');
+        }
         $subscription->delete();
     }
 }
